# src/main/java/com/cisco/ec2/model/GetEc2Response.java
package com.cisco.ec2.model;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;

import java.util.List;

@JsonInclude(JsonInclude.Include.NON_NULL)
@NoArgsConstructor
@AllArgsConstructor
@Builder
@Getter
public class GetEc2Response {

    private List<Ec2Instance> instances;
    private String nextToken;
}