# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8085",
  "basePath": "/",
...
  "paths": {
    "/aws-api/v1/regions/{region}/ec2": {
      "get": {
        "tags": [
          "ec-2-controller"
        ],
        "summary": "getEc2Instances",
        "operationId": "getEc2InstancesUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "nextToken",
            "in": "query",
            "description": "nextToken",
            "required": false,
            "type": "string"
          },
          {
            "name": "region",
            "in": "path",
            "description": "region",
            "required": true,
            "type": "string"
          },
          {
            "name": "size",
            "in": "query",
            "description": "max results per page",
            "required": false,
            "type": "integer",
            "format": "int32",
            "allowEmptyValue": false,
            "enum": [
              "5-500"
            ]
          },
          {
            "name": "sort",
            "in": "query",
            "description": "sort results by value",
            "required": false,
            "type": "string",
            "default": "instanceId",
            "allowEmptyValue": false,
            "enum": [
              "name",
              "instanceId",
              "state",
              "type",
              "availabilityZone",
              "privateIp",
              "publicIp"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetEc2Response"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
    "Ec2Instance": {
      "type": "object",
      "properties": {
        "availabilityZone": {
          "type": "string"
        },
        "instanceId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "privateIp": {
          "type": "string"
        },
        "publicIp": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      },
      "title": "Ec2Instance"
    },
    "GetEc2Response": {
      "type": "object",
      "properties": {
        "instances": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ec2Instance"
          }
        },
        "nextToken": {
          "type": "string"
        }
      },
      "title": "GetEc2Response"
    },
...
  }
}